Syncing with version 25.0.20074.0 (#26546)
Fixes
[AB#420000](https://dynamicssmb2.visualstudio.com/1fcb79e7-ab07-432a-a3c6-6cf5a88ba4a5/_workitems/edit/420000)

---------

+++ Apps/W1/Shopify/app/src/Order handling/Codeunits/ShpfyImportOrder.Codeunit.al
@@ -112,7 +112,8 @@ codeunit 30161 "Shpfy Import Order"
         OrderFulfillments.GetFulfillments(Shop, OrderHeader."Shopify Order Id");
 
         ConsiderRefundsInQuantityAndAmounts(OrderHeader);
-        DeleteZeroQuantityLines(OrderHeader);
+        if not Shop."Keep Zero Quantity Lines" then
+            DeleteZeroQuantityLines(OrderHeader);
 
         if CheckToCloseOrder(OrderHeader) then
             CloseOrder(OrderHeader);
